pyproject.toml
[project]
name = "document-qa_system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-huggingface>=0.3.1",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "tiktoken>=0.11.0",
]
